requirements.txt
flask>=2.3.3
flask-session>=0.5.0
flask-cors>=3.0.10
langchain>=0.1.0
langchain-core>=0.1.25
langgraph>=0.0.25
langchain-openai>=0.0.5
python-dotenv>=1.0.0 